requirements.txt
# ── Core ──
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# ── Web Scraping (Vulcan) ──
selenium>=4.8.3
beautifulsoup4>=4.12.0
webdriver-manager>=3.8.5
undetected-chromedriver>=3.5.0
requests>=2.28.2
feedparser>=6.0.10
numpy<2.0
pandas>=1.5.3,<2.3
fake-useragent>=0.1.11
lxml>=4.9.2

# ── RavenDB (Hades/SANS) ──
ravendb>=5.2.4
jsonpickle>=3.0.0

# ── Tor / Dark Web (nyx-crawler) ──
stem>=1.8.1
pysocks>=1.7.1
validators>=0.20.0
html2text>=2020.1.16

# ── Scheduler (Skadi) ──
apscheduler>=3.10.0
psycopg2-binary>=2.9.0

# ── OSINT Social ──
tweepy>=4.14.0
telethon>=1.34.0
asyncpraw>=7.7.0

# ── Public Records ──
python-whois>=0.9.4

# ── Storage ──
minio>=7.2.0

# ── Intelligence (Neo4j, Qdrant, Vision) ──
neo4j>=5.14.0
qdrant-client>=1.7.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
insightface>=0.7.3
onnxruntime>=1.16.0

# ── Speech / STT Forense (H4) ──
openai-whisper>=20231117
librosa>=0.10.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
gTTS>=2.3.0

# ── Logging ──
structlog>=23.2.0

# ── Crypto ──
cryptography>=40.0.1

# ── OSINT Enrichment ──
socid-extractor>=0.0.25
toutatis>=1.3
exifread>=3.0.0
